requirements.txt
requests
bs4
werkzeug==0.16.1
flask-session
cryptography<2.2
